Fix for Debian bug #108312
This fixes the "de" language, but doesn't fix other languages.
A proper detection (perhaps /[a-z]+ [0-9]+/) is more appropriate.

Of note also is that wu-ftpd had a bug filed against it
(which was fixed in 2003) which fixed this issue from the server
side - basically just always reporting "total", even if the
server wasn't in English. It was actually breaking a bunch of
other clients too.

Nonetheless, if this comes up again, at least we will know where
to look.