Fixed segfault if server disconnected (issue #19)
For sock_destroy to work properly, it needs to be followed
by a line which NULLs out the pointer, so there isn't a
double-free.

There was one place where this wasn't happening (probably not
the cause of the issue though), and one place where it was
happening, but AFTER the call to ftp_read_reply().

If ftp_read_reply() had an error (server closed control
connection), yafc would abort the connection, closing the
(already closed, but not marked as such) data connection,
causing a double-free.