Added stats after large PUTs (more than 20 files)
I don't know if this is the best way to do it or not,
but we can always change or revert it.

My thought was to also add in a config option (and command option?)
for setting the threshold.

Also, there might be a few bugs in the counts - I haven't
thouroughly tested recursive transfers (both -rf and -rn) for
example.